fix pyright

+++ src/toolregistry_hub/websearch/search_result.py
@@ -6,4 +6,4 @@ class SearchResult:
     title: str
     url: str
     content: str
-    score: int = 1.0
+    score: float = 1.0
